+++ src/dictionary/DefaultList.ts
@@ -4,18 +4,15 @@ import DictionaryValue from "../lib/DictionaryValue"
 export interface DefaultListInitOption extends DefaultModInitOption {
   align?: 'center' | 'left' | 'right'
   ellipsis?: boolean
-  auto?: boolean
 }
 
 class DefaultList extends DefaultMod {
   static $name = 'DefaultList'
   static $option = {
-    ellipsis: true,
-    auto: true
+    ellipsis: true
   }
   align: 'center' | 'left' | 'right'
   ellipsis: boolean
-  auto: boolean
   constructor(initOption: DefaultListInitOption | true, parent?: DictionaryValue, modName?: string) {
     if (initOption === true) {
       initOption = {}
@@ -24,7 +21,6 @@ class DefaultList extends DefaultMod {
     const $constructor = (this.constructor as typeof DefaultList)
     this.align = initOption.align || 'center'
     this.ellipsis = initOption.ellipsis === undefined ? $constructor.$option.ellipsis : initOption.ellipsis
-    this.auto = initOption.auto === undefined ? $constructor.$option.auto : initOption.auto
   }
 }